requirements.txt
flask
flask_socketio
socketio.client